requirements.txt
configargparse
python-dateutil
requests
tabulate
git+https://github.com/Rapptz/discord.py@rewrite